YTI-523: Convert deep concept search to use high level rest api. Terminology search remains tbd.

+++ src/main/java/fi/vm/yti/terminology/api/frontend/searchdto/DeepSearchConceptHitListDTO.java
@@ -7,7 +7,7 @@ public class DeepSearchConceptHitListDTO extends DeepSearchHitListDTO<ConceptSim
         super(Type.CONCEPT);
     }
 
-    public DeepSearchConceptHitListDTO(int totalCount, List<ConceptSimpleDTO> topHits) {
+    public DeepSearchConceptHitListDTO(long totalCount, List<ConceptSimpleDTO> topHits) {
         super(Type.CONCEPT, totalCount, topHits);
     }
 }
